Simplify providers by extracting behavior to other files
