+++ src/providers/jira/index.ts
@@ -1,10 +1,10 @@
-import { find, chain } from 'lodash';
+import { find } from 'lodash';
 
-import { Provider, PredicateType } from '../types';
-import { Filter } from '../../store/filters';
-import { makeSimplePredicate } from './predicates';
+import { Provider } from '../types';
+import { availablePredicates } from './predicates';
 import { Settings } from './components/Settings';
 import { IssueCard } from './components/IssueCard';
+import { fetchFilter } from './fetchers';
 
 class JiraProvider implements Provider {
   public id = 'jira';
@@ -12,43 +12,12 @@ class JiraProvider implements Provider {
   public settingsComponent = Settings;
   public cardComponent = IssueCard;
 
-  public getAvailablePredicates() {
-    return [
-      makeSimplePredicate({ name: 'project', placeholder: 'MYPROJECT' }),
-      makeSimplePredicate({ name: 'status', placeholder: 'open/closed' }),
-      makeSimplePredicate({ name: 'resolution', placeholder: '' }),
-      makeSimplePredicate({ name: 'sprint', placeholder: '' }),
-      makeSimplePredicate({ name: 'assignee', placeholder: '' }),
-    ];
-  }
+  public fetchFilter = fetchFilter;
+
+  public getAvailablePredicates = () => availablePredicates;
 
   public findPredicate(name: string) {
     return find(this.getAvailablePredicates(), { name });
-  }
-
-  public async fetchFilter(filter: Filter, settings: any): Promise<any[]> {
-    const site = '6345a5ad-6568-4d1a-8862-cfb76845117d'; // TODO
-    const token = settings.auth.access_token;
-
-    const filterString = chain(filter.predicates)
-      .map(predicate => filter.serializePredicate(predicate))
-      .map(encodeURIComponent)
-      .join('+AND+')
-      .value();
-
-    const url = `https://api.atlassian.com/ex/jira/${site}/rest/api/2/search?jql=${filterString}`;
-
-    const response = await fetch(url, {
-      method: 'GET',
-      headers: {
-        'Content-Type': 'application/json',
-        Authorization: `Bearer ${token}`,
-      },
-    });
-
-    const { issues } = await response.json();
-
-    return issues;
   }
 }
