+++ src/providers/jira/predicates/index.ts
@@ -23,3 +23,11 @@ export const makeSimplePredicate = ({
   type: PredicateType.TEXT,
   serialize: ({ value, negated }) => `${name} ${negated ? '!=' : '='} ${value}`,
 });
+
+export const availablePredicates: Predicate[] = [
+  makeSimplePredicate({ name: 'project', placeholder: 'MYPROJECT' }),
+  makeSimplePredicate({ name: 'status', placeholder: 'open/closed' }),
+  makeSimplePredicate({ name: 'resolution', placeholder: '' }),
+  makeSimplePredicate({ name: 'sprint', placeholder: '' }),
+  makeSimplePredicate({ name: 'assignee', placeholder: '' }),
+];
